requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/base.in
#
pygame==2.0.0             # via -r requirements/base.in
pyyaml==5.3.1             # via -r requirements/base.in
